Add files via upload
This program access a directory that is used to store feedback files in .txt format. 
For each file found it opens and process it into a Python dictionary, with appropriate keys such as 'title','name','date' and 'feedback'. 

It then post the data filled dictionary towards the server via http post request converting it into a json format and the response for that determines whether each file is successfully posted or not.